third_party/WebKit/LayoutTests/external/wpt/WebIDL/ecmascript-binding/sequence-conversion-expected.txt

This is a testharness.js-based test.
PASS An array
PASS A generator
FAIL An array with an overridden Symbol.iteratorassert_array_equals: property 0, expected 6 but got 1
FAIL An array with an overridden Symbol.iterator on the prototypeassert_array_equals: property 0, expected 11 but got 1
FAIL An array with an overridden %ArrayIterator%.prototype.nextassert_array_equals: property 0, expected 8 but got 1
PASS A holey array with fallback to an accessor on the prototype
Harness: the test ran to completion.
